Rename some methods
