+++ pkg/uninstall/uninstall.go
@@ -9,12 +9,12 @@ import (
 )
 
 func Uninstall(version string) {
-	if !paths.IsAlreadyDownloaded(version) {
+	if !paths.IsInstalled(version) {
 		err := fmt.Errorf("terraform version %s is not installed", helpers.ColoredVersion(version))
 		helpers.ExitWithError("uninstalling", err)
 	}
 
-	installLocation := paths.GetInstallLocation(version)
+	installLocation := paths.GetInstalledVersionLocation(version)
 	err := os.RemoveAll(installLocation)
 	if err != nil {
 		helpers.ExitWithError("deleting Terraform version", err)
