+++ pkg/use/use.go
@@ -21,7 +21,7 @@ func UseVersion(versionOrAlias string, autoInstall bool) {
 	}
 
 	// check if the version is installed
-	if !paths.IsAlreadyDownloaded(version) {
+	if !paths.IsInstalled(version) {
 		if !autoInstall {
 			err := fmt.Errorf("terraform version %s not found, run %s to install", helpers.ColoredVersion(version), helpers.ColoredInstallHelper(version))
 			helpers.ExitWithError("using", err)
